requirements.txt
alembic==1.14.1
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.3.0
Authlib==1.5.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
Bootstrap-Flask==2.4.1
boto3==1.38.22
botocore==1.38.22
browser-use==0.1.48
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
decompyle3==3.9.2
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
eventlet==0.39.1
executing==2.2.0
faiss-cpu==1.11.0
fastjsonschema==2.21.1
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fpdf==1.7.2
fsspec==2025.5.0
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
hyperframe==6.1.0
idna==3.8
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.42
Mako==1.3.8
markdownify==1.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mem0ai==0.1.93
mistune==3.1.3
monotonic==1.6
mpmath==1.3.0
mysqlclient==2.2.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numpy==2.2.6
oauthlib==3.2.2
ollama==0.4.8
openai==1.82.0
orjson==3.10.18
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.2.2
playwright==1.52.0
portalocker==2.10.1
posthog==3.25.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyMySQL==1.1.1
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2024.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.4.0
qdrant-client==1.14.2
qrcode==8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
screeninfo==0.8.1
sentence-transformers==4.1.0
setuptools==80.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
spark-parser==1.8.9
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
URLObject==2.4.3
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
xdis==6.1.5
yarg==0.1.9
zstandard==0.23.0